Ajout function JS updateURLParameter JS function et réorganisation JS files
