+++ static/js/mysite.js
@@ -1,85 +1,17 @@
+import MobileMenu from './mobileMenu.js';
+import WordCarousel from './wordCarousel.js';
+import StickyHeader from './stickyHeader.js';
+
 const MySite = {
     config: {
         carouselDelay: 3000,
         menuTransitionDelay: 300
     },
 
-    mobileMenu: {
-        init() {
-            const elements = ['burger-button', 'close-button', 'mobile-menu'].map(id => document.getElementById(id));
-            [this.burgerButton, this.closeButton, this.menu] = elements;
-            if (!this.menu) return;
-
-            this.panel = this.menu.querySelector('.fixed.inset-y-0');
-            this.overlay = this.menu.querySelector('.fixed.inset-0');
-            this.bindEvents();
-        },
-
-        bindEvents() {
-            const handlers = [
-                [this.burgerButton, () => this.open()],
-                [this.closeButton, () => this.close()],
-                [this.overlay, () => this.close()]
-            ];
-            handlers.forEach(([element, handler]) => element?.addEventListener('click', handler));
-        },
-
-        open() {
-            this.menu.classList.remove('hidden');
-            requestAnimationFrame(() => {
-                this.overlay.classList.add('opacity-100');
-                this.panel.classList.remove('translate-x-full');
-            });
-        },
-
-        close() {
-            this.overlay.classList.remove('opacity-100');
-            this.panel.classList.add('translate-x-full');
-            setTimeout(() => this.menu.classList.add('hidden'), MySite.config.menuTransitionDelay);
-        }
-    },
-
-    wordCarousel: {
-        init() {
-            this.words = document.querySelectorAll('.rotating-text');
-            if (this.words.length === 0) return;
-            this.currentIndex = 0;
-            this.startRotation();
-        },
-
-        startRotation() {
-            setInterval(() => {
-                const current = this.words[this.currentIndex];
-                const next = this.words[(this.currentIndex + 1) % this.words.length];
-                
-                current.classList.remove('active');
-                current.classList.add('hidden');
-                next.classList.remove('hidden');
-                next.classList.add('active');
-                
-                this.currentIndex = (this.currentIndex + 1) % this.words.length;
-            }, MySite.config.carouselDelay);
-        }
-    },
-
-    stickyHeader: {
-        init() {
-            const header = document.querySelector('[data-sticky-header]');
-            if (!header) return;
-
-            let lastScroll = 0;
-            window.addEventListener('scroll', () => {
-                const currentScroll = window.scrollY;
-                header.classList.toggle('shadow-lg', currentScroll > 0);
-                lastScroll = currentScroll;
-            });
-        }
-    },
-    
     init() {
-        this.mobileMenu.init();
-        this.wordCarousel.init();
-        this.stickyHeader.init();
+        MobileMenu.init();
+        WordCarousel.init();
+        StickyHeader.init();
     }
 };
 
@@ -115,3 +47,9 @@ function toggleSubMenu(button) {
         svg.classList.add('rotate-180');
     }
 }
+
+window.updateURLParameter = function(param, value) {
+    const url = new URL(window.location.href);
+    url.searchParams.set(param, value);
+    window.location.href = url.toString();
+};
